Fix inverted predecessor/successor in relations tab
A relation should be read as `<from> <relation_type> <to>`: for instance
`Relation.create(from_id: 123, to_id: 456, relation_type: :follows)`
means that work package 123 follows work package 456, which means that
work package 123 is the successor of work package 456, or that work
package 456 is the predecessor of work package 123.

The translations were inverted.

Also used `shared_let` in specs to speed them up a little.