Revert "fix[Op#59433]: simplify reminder model to directly linked to notification"
This reverts commit ae3f77da803f3eac97e5e12121f1266f807b7d31.

Maintain a persistent record of reminder notifications as notifications are always retrievable by the user, e.g. via
email, viewing "all" notifications.

A reminder notification is only created IF there are no active unread notifications, hence when rescheduling, snoozing or editing a reminder then
ensure all previous notifications are marked as read- as we should not have duplicate notifications for a given reminder.