[#55581] make hover card work with avatar fallbacks
This was a hard one to debug. When you hover over an avatar as soon as
the page loads, you will hover over the fallback first. Within the
timeout, the actual avatar image loads and replaces the fallback. After
that, the timeout of the hover trigger runs out and displays the hover
cards, but does no longer have an element to position relative to. So,
we check for the existence of the hovered element before showing a hover
card to fix this.