Beautify dev highlight
This might be an issue dependent on the OS or
available fonts, but to me the header in development mode
read like "developmerdevelopmer", because the text was wider
than the image width we allocated.

By increasing the width to 120px, there is now some whitespace
left between the images (hopefully for most OSes), which looks
nicer IMO and makes the dev-hint more readable.