Make it possible to use workers in TLS setup
The example overrides file was missing overrides
for the workers, which means their network was not
configured to include the gateway network, thus they
didn't resolve IP addresses to the Traefik proxy.

Even if they had, it was also necessary to setup environment
variables the same way that the backend service gets them, because
otherwise the development TLS certificates would not be recognized
and configuration of KeyCloak would not be taken into account for
workers, which may be an issue in the future.